refactor(ui): Make draggable faster by changing how you configure move()
Now move is the return value of start. If this wasn't like this, then variables to be reached in move (in Deriv) would have to be in the component, but only one component's draggable listeners are active at a time. So all the other Deriv components would have some unused variable all the time...