feat(ui): Readd the clip-out feature and finally implement correctly (?)
Had a lot of trouble with this one. Now when the dragged deriv is transitioning out of a zone but not entering any, it has a clip behaviour where it gets pushed around by the zones next to it. Just clipToBoundingRect by itself without clip-out was unusable. (Can't move derivs to some regions because it keeps getting teleported away.))