refactor(state): Transfer logic files from pl-proto (except Label)
Logic stuff arent't yet implemented but the transfered code is changed so that: it doesn't show any errors and comments and variable names are consistent with the current project. Error isn'r a separate file anymore. And label will be implemented from scratch using the new Crawler API. I'm not happy with the current implementation of logic but I'll keep it for now and revamp later when adding 1st order logic. (Even tho the current impl. can also handle FOL.)