feat(ui): Anchor API (+ xTransform -> xTranslate)
This was needed to fix an issue where whıle doing DND, the dragged element jumps around when some rule's are recalculated. This seems to be the most obvious solution. And I may add another anchor later, but even in this case it seems a bit better than before despite there being only 2 anchors.