refactor(ui): Refactor constants to avoid recalculation
Putting (non-constant expression)+(constant expression) in svelte reactive attribute fields causes the constant expressions to be recalculated everytime with the variable part. So I defined constants for those parts to maybe make it perform better? This may be a bit of a bad-practive but I'm not sure. Will keep it this way for now.