refactor(ui): Carry viewport event listeners to the sorrounding div
Since the bind:this can't be put in panzoom, it was put in the srrounding div. I'm carrying the other listeners to be consistent with that. (I can expose panzoom's element as a prop then continue using the panzoom element but this overall made more sense, it's a bit simpler?)