refactor: Part + UID System
Now every element of viewport will be in some part (typed string). Hover and selection is to be tracked in viewport, now most interactions and hover effects will be scheduled in viewport too.  This will help avoid having independent listeners all over the place interacting with each other.  (Even now the DND listeners glitch the selection system.) For more information see the comments in the deriv render data file. Also the uid.null is gone since it was solely for the old just-UID system. And handle component doesn't have a data-part because it will be removed. Deriv DND will be managed in viewport with draggable.once.