refactor(ui): Add new hover state where deriv's hovered but not body or bar
This is to support usage of native hover alongside them. This sort of hybrid system will be used for adders. The non-native hover system was made only to style Bg. Expanding it to handle adders isn't reasonable. Although I implemented that in a previouus commit and reverted it. See dae857a114d9d57a2ea55e2a68cebec6909ca379.