refactor: rafPromise -> tick
The thing the code is intended to await when await rafPromise() is called is for html to be updated with svelte states. There's already tick for that, which I had forgotten about. (raf: requestAnimationFrame) Using raf might have even allowed some bugs where if the program is laggy enough, tick should finish way after rafPromise (I don't know if this is actually possible)...