refactor(ui): Add -1 tab index to panzoom and deriv
These are interactable elements that should be able to recieve focus. Thid will also be used to implement the deselect-on-blur behavior of MultiSelection in the next commits. (Panzoom won't cause a deselect, but for this exception case to be detectable in the blur event, panzoom should be able to recieve focus. Deriv also has to bc otherwise clicking on it will count as clicking on panzoom...)